[core] use internal `/run` for temp files
Added it to pidfile (instead of `/var/run/dd-agent`) and pickle
files, as it should fix all issues with `dd-agent info` reporting
incorrectly. It doesn't break anything, as it checks first to see if
the internal `/run` exists, and fallback to the current behaviour if
not. (ie `tempfile.gettempdir()`)

Also:
- move PidFile and Platform from util.py to their own files in `utils/`
- as a consequence, move `checks/utils.py` into `utils/tailfile.py`
