+++ checks/system/unix.py
@@ -15,7 +15,8 @@ import time
 
 # project
 from checks import Check, UnknownValue
-from util import get_hostname, Platform
+from util import get_hostname
+from utils.platform import Platform
 
 # locale-resilient float converter
 to_float = lambda s: float(s.replace(",", "."))
