+++ packaging/datadog-agent/source/supervisord.conf
@@ -5,18 +5,17 @@ loglevel = info
 nodaemon = true
 identifier = supervisor
 nocleanup = true
-pidfile = supervisord/supervisord.pid
-directory= supervisord
+pidfile = %(here)s/../run/supervisord.pid
 
 [rpcinterface:supervisor]
 supervisor.rpcinterface_factory = supervisor.rpcinterface:make_main_rpcinterface
 
 [unix_http_server]
-file = supervisord/agent-supervisor.sock
+file = %(here)s/../run/agent-supervisor.sock
 
 [supervisorctl]
 prompt = datadog
-serverurl = unix://supervisord/agent-supervisor.sock
+serverurl = unix://%(here)s/../run/agent-supervisor.sock
 
 [program:collector]
 command=python agent/agent.py foreground --use-local-forwarder
