+++ packaging/debian/datadog-agent.init
@@ -20,9 +20,9 @@ AGENTUSER="dd-agent"
 FORWARDERPATH="/opt/datadog-agent/agent/ddagent.py"
 NAME="datadog-agent"
 DESC="Datadog Agent"
-SUPERVISOR_PIDFILE="/var/run/datadog-supervisord.pid"
+SUPERVISOR_PIDFILE="/opt/datadog-agent/run/datadog-supervisord.pid"
 SUPERVISOR_FILE="/etc/dd-agent/supervisor.conf"
-SUPERVISOR_SOCK="/var/tmp/datadog-supervisor.sock"
+SUPERVISOR_SOCK="/opt/datadog-agent/run/datadog-supervisor.sock"
 SUPERVISORCTL_PATH="/opt/datadog-agent/bin/supervisorctl"
 SUPERVISORD_PATH="/opt/datadog-agent/bin/supervisord"
 SYSTEM_PATH=/opt/datadog-agent/embedded/bin:/opt/datadog-agent/bin:$PATH
