+++ packaging/supervisor.conf
@@ -1,21 +1,21 @@
 [supervisorctl]
-serverurl = unix:///var/tmp/datadog-supervisor.sock
+serverurl = unix:///opt/datadog-agent/run/datadog-supervisor.sock
 
 [unix_http_server]
-file=/var/tmp/datadog-supervisor.sock
+file=/opt/datadog-agent/run/datadog-supervisor.sock
 
 [rpcinterface:supervisor]
 supervisor.rpcinterface_factory = supervisor.rpcinterface:make_main_rpcinterface
 
 [supervisord]
-http_port = /var/tmp/datadog-supervisor.sock
+http_port = /opt/datadog-agent/run/datadog-supervisor.sock
 minfds = 1024
 minprocs = 200
 loglevel = info
 logfile = /var/log/datadog/supervisord.log
 logfile_maxbytes = 50MB
 nodaemon = false
-pidfile = /var/run/datadog-supervisord.pid
+pidfile = /opt/datadog-agent/run/datadog-supervisord.pid
 logfile_backups = 10
 environment=PYTHONPATH=/opt/datadog-agent/agent:/opt/datadog-agent/agent/checks,LANG=POSIX
