+++ tests/checks/integration/test_tomcat.py
@@ -12,7 +12,7 @@ from nose.plugins.attrib import attr
 from aggregator import MetricsAggregator
 from dogstatsd import Server
 from jmxfetch import JMXFetch
-from util import PidFile
+from utils.pidfile import PidFile
 
 STATSD_PORT = 8126
 
@@ -64,7 +64,7 @@ class TestTomcat(unittest.TestCase):
         while self.reporter.metrics is None:
             time.sleep(1)
             count += 1
-            if count > 25:
+            if count > 25:ga
                 raise Exception("No metrics were received in 25 seconds")
 
         metrics = self.reporter.metrics
