+++ tests/checks/mock/test_solr.py
@@ -12,7 +12,7 @@ from aggregator import MetricsAggregator
 from dogstatsd import Server
 from jmxfetch import JMXFetch
 from tests.checks.common import Fixtures
-from util import PidFile
+from utils.pidfile import PidFile
 
 STATSD_PORT = 8127
