+++ tests/checks/mock/test_system_core.py
@@ -1,8 +1,8 @@
 import mock
 import psutil
 
-from tests.checks.common import AgentCheckTest
-from util import Platform
+from tests.common import AgentCheckTest
+from utils.platform import Platform
 
 if Platform.is_mac():
     CHECK_RATES = [
