+++ tests/core/test_config.py
@@ -7,7 +7,9 @@ import unittest
 
 # project
 from config import get_config, load_check_directory, DEFAULT_CHECKS
-from util import PidFile, is_valid_hostname, Platform, windows_friendly_colon_split
+from util import is_valid_hostname, windows_friendly_colon_split
+from utils.pidfile import PidFile
+from utils.platform import Platform
 
 
 class TestConfig(unittest.TestCase):
